Explicitly set DirectoryCopyFlags to "timestamps".
This parameter has a different default setting with RoboCopy in the
Server 2012 version: "DA" (data and attributes), compared to "T"
(timestamps) on e.g. Windows 7.

Forcing this setting to be the same for 2012 and 7 seems to fix the
issue of files not being transferred. More testing is needed though!

Refers to #11